requirements.txt
# ============================================================
# Financial Document Analyzer — Python Dependencies
# ============================================================
# Keep crewai at version 0.130.0
# Other package versions are flexible and can be changed
# Only change crewai version if there are critical dependency
# conflicts that cannot be resolved by other means
# ============================================================

# --- Core Framework ---
crewai==0.130.0
crewai-tools==0.47.1

# --- Web Framework ---
fastapi==0.110.3
uvicorn>=0.29.0
python-multipart>=0.0.9

# --- LLM Providers ---
google-generativeai>=0.5.4
google-ai-generativelanguage>=0.6.4
google-api-core>=2.10.0
google-api-python-client>=2.131.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
openai>=1.30.5

# --- PDF Processing ---
langchain-community>=0.0.38
pypdf>=4.0.0

# --- LangChain Core ---
langchain-core>=0.1.52
langsmith>=0.1.67

# --- Environment ---
python-dotenv>=1.0.0

# --- Database (Bonus Feature) ---
sqlalchemy>=2.0.0

# --- Queue Worker (Bonus Feature) ---
celery>=5.3.0
redis>=5.0.0

# --- Data Processing ---
numpy>=1.26.4
pandas>=2.2.2
pydantic>=1.10.13

# --- Telemetry (CrewAI dependency) ---
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-http>=1.25.0

# --- Other ---
click>=8.1.7
Jinja2>=3.1.4
jsonschema>=4.22.0
protobuf>=4.25.3
pillow>=10.3.0